added JavaDoc and assert to use environment

+++ org.kompiro.jamcircle.storage/src/org/kompiro/jamcircle/storage/service/internal/FileStorageServiceImpl.java
@@ -9,7 +9,7 @@ import org.apache.commons.io.FileUtils;
 import org.eclipse.core.runtime.Platform;
 import org.kompiro.jamcircle.storage.Messages;
 import org.kompiro.jamcircle.storage.StorageStatusHandler;
-import org.kompiro.jamcircle.storage.exception.FileException;
+import org.kompiro.jamcircle.storage.exception.FileStorageException;
 import org.kompiro.jamcircle.storage.service.FileStorageService;
 
 public class FileStorageServiceImpl implements FileStorageService {
@@ -80,12 +80,15 @@ public class FileStorageServiceImpl implements FileStorageService {
 		return new File(storeRoot).getAbsolutePath() + File.separator;
 	}
 	
-	public void deleteAll() {
+	public void deleteAll()  throws FileStorageException{
+		if(testmode == false){
+			throw new FileStorageException("This API might call cleaning for test only.");
+		}
 		File root = new File(getStoreRoot());
 		try {
 			FileUtils.deleteDirectory(root);
 		} catch (IOException e) {
-			throw new FileException(e);
+			throw new FileStorageException(e);
 		}
 	};
